requirements.txt
aiogram==2.25.1
gspread>=5.7.0
google-auth>=2.20.0
httpx>=0.24.0
